Factored out colour and plain-text components in bash prompt
This is to make way for doing layout changes based on text length.
I had to pull in the git_prompt function into the main prompt
function because bash sucks at returning multiple values, even
arrays. I made a nice text work-around by breaking the string with a \n,
but it was overly complicated.